backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
biopython>=1.83
numpy>=1.26
pydantic>=2.0